Move utils/find_shadowed_files.py to ssg/_utils_shadowed.py
